feat: add stop button
